requirements.txt
requests
python-dotenv
google-cloud-bigquery
db-dtypes
pandas
tabulate
